+++ src/compiler/output-targets/test/output-targets-dist.spec.ts
@@ -1,9 +1,10 @@
-import { doNotExpectFiles, expectFiles } from '../../../testing/testing-utils';
+// @ts-nocheck
+import { expectFilesDoNotExist, expectFilesExist } from '../../../testing/testing-utils';
 import { Compiler, Config } from '@stencil/core/compiler';
 import { mockConfig } from '@stencil/core/testing';
 import path from 'path';
 
-xdescribe('outputTarget, dist', () => {
+describe.skip('outputTarget, dist', () => {
   jest.setTimeout(20000);
   let compiler: Compiler;
   let config: Config;
@@ -52,7 +53,7 @@ xdescribe('outputTarget, dist', () => {
     const r = await compiler.build();
     expect(r.diagnostics).toHaveLength(0);
 
-    expectFiles(compiler.fs, [
+    expectFilesExist(compiler.fs, [
       path.join(root, 'User', 'testing', 'dist', 'index.js'),
       path.join(root, 'User', 'testing', 'dist', 'index.mjs'),
       path.join(root, 'User', 'testing', 'dist', 'index.js.map'),
@@ -81,7 +82,7 @@ xdescribe('outputTarget, dist', () => {
       path.join(root, 'User', 'testing', 'src', 'components.d.ts'),
     ]);
 
-    doNotExpectFiles(compiler.fs, [
+    expectFilesDoNotExist(compiler.fs, [
       path.join(root, 'User', 'testing', 'build'),
       path.join(root, 'User', 'testing', 'esm'),
       path.join(root, 'User', 'testing', 'es5'),