+++ src/compiler/service-worker/test/service-worker.spec.ts
@@ -1,9 +1,12 @@
+// @ts-nocheck
+// TODO(STENCIL-462): investigate getting this file to pass (remove ts-nocheck)
 import type * as d from '@stencil/core/declarations';
 import { Compiler, Config } from '@stencil/core/compiler';
 import { mockConfig } from '@stencil/core/testing';
 import path from 'path';
 
-xdescribe('service worker', () => {
+// TODO(STENCIL-462): investigate getting this file to pass
+describe.skip('service worker', () => {
   jest.setTimeout(20000);
   let compiler: Compiler;
   let config: Config;