+++ src/mock-doc/node.ts
@@ -4,7 +4,14 @@ import { matches, selectAll, selectOne } from './selector';
 import { MockAttr, MockAttributeMap, createAttributeProxy } from './attribute';
 import { MockClassList } from './class-list';
 import { MockCSSStyleDeclaration, createCSSStyleDeclaration } from './css-style-declaration';
-import { MockEvent, addEventListener, dispatchEvent, removeEventListener, resetEventListeners } from './event';
+import {
+  MockEvent,
+  addEventListener,
+  dispatchEvent,
+  removeEventListener,
+  resetEventListeners,
+  MockFocusEvent,
+} from './event';
 import { NODE_NAMES, NODE_TYPES } from './constants';
 import { NON_ESCAPABLE_CONTENT, SerializeNodeToHtmlOptions, serializeNodeToHtml } from './serialize-node';
 import { parseFragmentUtil } from './parse-util';
@@ -141,7 +148,12 @@ export class MockNode {
     if (otherNode === this) {
       return true;
     }
-    return this.childNodes.includes(otherNode);
+    const childNodes = Array.from(this.childNodes);
+    if (childNodes.includes(otherNode)) {
+      return true;
+    }
+
+    return childNodes.some((node) => this.contains.bind(node)(otherNode));
   }
 
   removeChild(childNode: MockNode) {
@@ -230,6 +242,13 @@ export class MockElement extends MockNode {
     return shadowRoot;
   }
 
+  blur() {
+    dispatchEvent(
+      this,
+      new MockFocusEvent('blur', { relatedTarget: null, bubbles: true, cancelable: true, composed: true })
+    );
+  }
+
   get shadowRoot() {
     return this.__shadowRoot || null;
   }
@@ -309,6 +328,13 @@ export class MockElement extends MockNode {
 
   get firstElementChild(): MockElement | null {
     return this.children[0] || null;
+  }
+
+  focus(_options?: { preventScroll?: boolean }) {
+    dispatchEvent(
+      this,
+      new MockFocusEvent('focus', { relatedTarget: null, bubbles: true, cancelable: true, composed: true })
+    );
   }
 
   getAttribute(attrName: string) {